Long instead of Int

+++ src/main/kotlin/dk/fantastiskefroe/it/shopify_data/dto/webhook/RefundLineItemDTO.kt
@@ -1,6 +1,6 @@
 package dk.fantastiskefroe.it.shopify_data.dto.webhook
 
 data class RefundLineItemDTO(
-    val lineItemId: Int,
+    val lineItemId: Long,
     val restockType: RestockTypeDTO
 )
